Move key input handlers to separate folder
