+++ menu/CMakeLists.txt
@@ -68,10 +68,10 @@ add_library(${PROJECT_NAME} MODULE src/dllmain.cpp
         src/ui/components/component_flags.hpp
         src/ui/components/execute_component.hpp
         src/ui/components/sub_link_component.hpp
-        src/util/key_event_listener.hpp
-        src/util/key_event_listener.cpp
-        src/util/key_state.cpp
-        src/util/key_state.hpp
+        src/util/key_input/key_event_listener.hpp
+        src/util/key_input/key_event_listener.cpp
+        src/util/key_input/key_state.cpp
+        src/util/key_input/key_state.hpp
 )
 target_compile_definitions(${PROJECT_NAME} PUBLIC BASE_SUBCOMPONENT="${PROJECT_NAME}")
